feat: improved tests

+++ scripts/api.py
@@ -26,7 +26,7 @@ def dummy_function(*args, **kwargs):
     return None
 
 
-logger = logging.getLogger(__name__)
+logger = logging.getLogger("sd_proxy")
 logger.setLevel(os.environ.get('LOG_LEVEL') or logging.ERROR)
 
 
@@ -165,6 +165,8 @@ def parse_constant(c: str) -> float:
 
 
 def sagemaker_api(_, app: FastAPI):
+    logger.info(app.__dict__)
+    logger.info(app)
     logger.debug("Loading Sagemaker API Endpoints.")
     import threading
     from collections import deque
@@ -175,11 +177,8 @@ def sagemaker_api(_, app: FastAPI):
 
     @app.post("/invocations")
     def invocations(req: InvocationsRequest):
-        """
-        Check the current state of Dreambooth processes.
-        @return:
-        """
-        logger.info('-------invocation------')
+
+        logger.info(f'-------invocation on port {req.port}------')
         logger.info(json.dumps(req.__dict__, default=str))
 
         with condition:
@@ -215,7 +214,7 @@ def sagemaker_api(_, app: FastAPI):
                     if image_type:
                         logger.debug(f"set output_img_type:{image_type}")
                         resp["output_img_type"] = image_type
-                    response = requests.post(url=f'http://0.0.0.0:8080/sdapi/v1/txt2img',
+                    response = requests.post(url=f'http://0.0.0.0:{req.port}/sdapi/v1/txt2img',
                                              json=payload)
                     logger.info(f"{threading.current_thread().ident}_{threading.current_thread().name}_______ txt2img end !!!!!!!! {len(response.json())}")
                     resp.update(response.json())
@@ -230,25 +229,25 @@ def sagemaker_api(_, app: FastAPI):
                     if image_type:
                         logger.debug(f"set output_img_type:{image_type}")
                         resp["output_img_type"] = image_type
-                    response = requests.post(url=f'http://0.0.0.0:8080/sdapi/v1/img2img',
+                    response = requests.post(url=f'http://0.0.0.0:{req.port}/sdapi/v1/img2img',
                                              json=payload)
                     logger.info(f"{threading.current_thread().ident}_{threading.current_thread().name}_______ img2img end !!!!!!!!{len(response.json())}")
                     resp.update(response.json())
                     return resp
                 elif req.task == 'interrogate_clip' or req.task == 'interrogate_deepbooru':
-                    response = requests.post(url=f'http://0.0.0.0:8080/sdapi/v1/interrogate',
+                    response = requests.post(url=f'http://0.0.0.0:{req.port}/sdapi/v1/interrogate',
                                              json=json.loads(req.interrogate_payload.json()))
                     return response.json()
                 elif req.task == 'extra-single-image':
-                    response = requests.post(url=f'http://0.0.0.0:8080/sdapi/v1/extra-single-image',
+                    response = requests.post(url=f'http://0.0.0.0:{req.port}/sdapi/v1/extra-single-image',
                                              json=payload)
                     return response.json()
                 elif req.task == 'extra-batch-images':
-                    response = requests.post(url=f'http://0.0.0.0:8080/sdapi/v1/extra-batch-images',
+                    response = requests.post(url=f'http://0.0.0.0:{req.port}/sdapi/v1/extra-batch-images',
                                              json=payload)
                     return response.json()
                 elif req.task == 'rembg':
-                    response = requests.post(url=f'http://0.0.0.0:8080/rembg', json=payload)
+                    response = requests.post(url=f'http://0.0.0.0:{req.port}/rembg', json=payload)
                     return response.json()
                 elif req.task == 'db-create-model':
                     r"""
@@ -292,7 +291,7 @@ def sagemaker_api(_, app: FastAPI):
                                 input_path = os.path.join(get_path_from_s3_path(s3_input_path), local_model_path)
                                 logger.info(f"ckpt from local {input_path} {local_model_path}")
                             input_bucket_name = get_bucket_name_from_s3_path(s3_input_path)
-                            logging.info("Check disk usage before download.")
+                            logger.info("Check disk usage before download.")
                             os.system("df -h")
                             logger.info(f"Download src model from s3 {input_bucket_name} {input_path} {local_model_path}")
                             download_folder_from_s3_by_tar(input_bucket_name, input_path, local_model_path)
@@ -301,14 +300,10 @@ def sagemaker_api(_, app: FastAPI):
                             logger.info("Check disk usage after download.")
                             os.system("df -h")
                         logger.info("Start creating model.")
-                        # local_response = requests.post(url=f'http://0.0.0.0:8080/dreambooth/createModel',
-                        #                         params=db_create_model_params)
                         create_model_func_args = copy.deepcopy(db_create_model_params)
-                        # ckpt_path = create_model_func_args.pop("new_model_src")
-                        # create_model_func_args["ckpt_path"] = ckpt_path
                         local_response = create_model(**create_model_func_args)
                         target_local_model_dir = f'models/dreambooth/{db_create_model_params["new_model_name"]}'
-                        logging.info(f"Upload tgt model to s3 {target_local_model_dir} {output_bucket_name} {output_path}")
+                        logger.info(f"Upload tgt model to s3 {target_local_model_dir} {output_bucket_name} {output_path}")
                         upload_folder_to_s3_by_tar(target_local_model_dir, output_bucket_name, output_path)
                         config_file = os.path.join(target_local_model_dir, "db_config.json")
                         with open(config_file, 'r') as openfile:
@@ -338,11 +333,11 @@ def sagemaker_api(_, app: FastAPI):
                         delete_src_command = f"rm -rf models/Stable-diffusion/{db_create_model_params['ckpt_path']}"
                         logger.info(delete_src_command)
                         os.system(delete_src_command)
-                        logging.info("Delete tgt model.")
+                        logger.info("Delete tgt model.")
                         delete_tgt_command = f"rm -rf models/dreambooth/{db_create_model_params['new_model_name']}"
                         logger.info(delete_tgt_command)
                         os.system(delete_tgt_command)
-                        logging.info("Check disk usage after request.")
+                        logger.info("Check disk usage after request.")
                         os.system("df -h")
                 elif req.task == 'merge-checkpoint':
                     try:
@@ -384,19 +379,17 @@ def get_file_md5_dict(path):
 
 
 def move_model_to_tmp(_, app: FastAPI):
-    # os.system("rm -rf models")
-    # Create model dir
-    # logger.info("Create model dir")
-    # os.system("mkdir models")
-    # Move model dir to /tmp
-    logging.info("Copy model dir to tmp")
-    model_tmp_dir = f"models_{time.time()}"
+    logger.info("Copy model dir to tmp")
+    model_tmp_dir = f"models"
+    # for mutil gpus
+    if os.path.exists(f"/tmp/{model_tmp_dir}"):
+        return
     os.system(f"cp -rL models /tmp/{model_tmp_dir}")
     src_file_dict = get_file_md5_dict("models")
     tgt_file_dict = get_file_md5_dict(f"/tmp/{model_tmp_dir}")
     is_complete = True
     for file in src_file_dict:
-        logging.info(f"Src file {file} md5 {src_file_dict[file]}")
+        logger.info(f"Src file {file} md5 {src_file_dict[file]}")
         if file not in tgt_file_dict:
             is_complete = False
             break
@@ -405,15 +398,11 @@ def move_model_to_tmp(_, app: FastAPI):
             break
     if is_complete:
         os.system(f"rm -rf models")
-        # Delete tmp model dir
-        # logger.info("Delete tmp model dir")
-        # os.system("rm -rf /tmp/models")
-        # Link model dir
-        logging.info("Link model dir")
+        logger.info("Link model dir")
         os.system(f"ln -s /tmp/{model_tmp_dir} models")
     else:
-        logging.info("Failed to copy model dir, use the original dir")
-    logging.info("Check disk usage on app started")
+        logger.info("Failed to copy model dir, use the original dir")
+    logger.info("Check disk usage on app started")
     os.system("df -h")
 
 try:
